Separate responsabilities
