+++ Editor/CustomEditors/AnimationClipAssetEditor.cs
@@ -2,9 +2,9 @@ using DOTSAnimation.Authoring;
 using UnityEditor;
 using UnityEditor.UIElements;
 using UnityEngine;
-using UnityEngine.Animations;
-using UnityEngine.Playables;
+using UnityEngine.Assertions;
 using UnityEngine.UIElements;
+using Object = UnityEngine.Object;
 
 namespace DOTSAnimation.Editor
 {
@@ -12,17 +12,13 @@ namespace DOTSAnimation.Editor
     public class AnimationClipAssetEditor : UnityEditor.Editor
     {
         public VisualTreeAsset DefaultInspector;
-        private PreviewRenderUtility previewRenderUtility;
-        private GameObject gameObject;
-        private Animator animator;
-        private SkinnedMeshRenderer skinnedMeshRenderer;
-        private Mesh previewMesh;
-        private PlayableGraph playableGraph;
-        private float sampleNormalizedTime;
+        private SingleClipPreview preview;
         private AnimationClipAsset ClipTarget => (AnimationClipAsset)target;
 
         public override VisualElement CreateInspectorGUI()
         {
+            Assert.IsNotNull(preview);
+            
             var inspector = new VisualElement();
             if (DefaultInspector != null)
             {
@@ -44,13 +40,13 @@ namespace DOTSAnimation.Editor
             if (slider != null)
             {
                 slider.RegisterValueChangedCallback(OnTimeSliderValueChanged);
-                slider.value = sampleNormalizedTime;
+                slider.value = preview.SampleNormalizedTime;
             }
 
             var previewObjField = inspector.Q<ObjectField>("preview-obj");
             if (previewObjField != null)
             {
-                previewObjField.value = gameObject;
+                previewObjField.value = preview.GameObject;
                 previewObjField.RegisterValueChangedCallback(OnPreviewObjectChanged);
             }
             return inspector;
@@ -58,155 +54,28 @@ namespace DOTSAnimation.Editor
 
         private void OnPreviewObjectChanged(ChangeEvent<Object> evt)
         {
-            gameObject = (GameObject) evt.newValue;
-            if (gameObject != null)
-            {
-                if (!TryInstantiateSkinnedMesh(gameObject))
-                {
-                    DestroyPreviewInstance();
-                    gameObject = null;
-                }
-            }
-            else
-            {
-                DestroyPreviewInstance();
-            }
+            preview.GameObject = (GameObject) evt.newValue;
         }
 
         private void OnObjectPropertyChanged(ChangeEvent<Object> evt)
         {
-            RefreshPreviewObjects();
+            preview.Clip = ClipTarget.Clip;
         }
 
         private void OnTimeSliderValueChanged(ChangeEvent<float> evt)
         {
-            sampleNormalizedTime = evt.newValue;
-        }
-
-        private bool IsValidGameObject(GameObject obj)
-        {
-            return obj.GetComponentInChildren<Animator>() != null;
-        }
-
-        private bool TryInstantiateSkinnedMesh(GameObject template)
-        {
-            if (!IsValidGameObject(template))
-            {
-                return false;
-            }
-
-            DestroyPreviewInstance();
-
-            var instance = Instantiate(template, Vector3.zero, Quaternion.identity);
-            
-            AnimatorUtility.DeoptimizeTransformHierarchy(instance);
-            
-            instance.hideFlags = HideFlags.DontSaveInEditor | HideFlags.DontSaveInBuild | HideFlags.HideInHierarchy |
-                                 HideFlags.HideInInspector | HideFlags.NotEditable;
-            animator = instance.GetComponentInChildren<Animator>();
-            skinnedMeshRenderer = instance.GetComponentInChildren<SkinnedMeshRenderer>();
-            instance.SetActive(false);
-
-            previewMesh = new Mesh();
-
-            if (playableGraph.IsValid())
-            {
-                playableGraph.Destroy();
-            }
-            playableGraph = PlayableGraph.Create();
-            playableGraph.SetTimeUpdateMode(DirectorUpdateMode.Manual);
-            var playableOutput = AnimationPlayableOutput.Create(playableGraph, "Animation", animator);
-            var clipPlayable = AnimationClipPlayable.Create(playableGraph, ClipTarget.Clip);
-            playableOutput.SetSourcePlayable(clipPlayable);
-            
-            CreatePreviewUtility();
-
-            return true;
-        }
-
-        private void DestroyPreviewInstance()
-        {
-            if (skinnedMeshRenderer != null)
-            {
-                DestroyImmediate(skinnedMeshRenderer.transform.root.gameObject);
-                skinnedMeshRenderer = null;
-                DestroyImmediate(previewMesh);
-                previewMesh = null;
-            }
-        }
-
-        private void CreatePreviewUtility()
-        {
-            previewRenderUtility?.Cleanup();
-            previewRenderUtility = new PreviewRenderUtility();
-
-            var bounds = skinnedMeshRenderer.bounds;
-            var camPos = new Vector3(0f, bounds.size.y * 3, 10f);
-            previewRenderUtility.camera.transform.position = camPos;
-            var camRot = Quaternion.LookRotation(bounds.center - camPos);
-            previewRenderUtility.camera.transform.rotation = camRot;
-            previewRenderUtility.camera.nearClipPlane = 0.3f;
-            previewRenderUtility.camera.farClipPlane = 3000f;
-
-            var light1 = previewRenderUtility.lights[0];
-            light1.type = LightType.Directional;
-            light1.color = Color.white;
-            light1.intensity = 1;
-            light1.transform.rotation = camRot;
+            preview.SampleNormalizedTime = evt.newValue;
         }
 
         private void OnEnable()
         {
-            AnimationMode.StartAnimationMode();
-            RefreshPreviewObjects();
+            preview = new SingleClipPreview(ClipTarget.Clip);
+            preview.Initialize();
         }
         
         private void OnDisable()
         {
-            AnimationMode.StopAnimationMode();
-            previewRenderUtility?.Cleanup();
-        }
-
-
-        private void RefreshPreviewObjects()
-        {
-            if (gameObject != null)
-            {
-                if (!TryInstantiateSkinnedMesh(gameObject))
-                {
-                    gameObject = null;
-                }
-            }
-            else if (ClipTarget.Clip != null)
-            {
-                if (TryFindSkeletonFromClip(ClipTarget.Clip, out var armatureGo))
-                {
-                    if (TryInstantiateSkinnedMesh(armatureGo))
-                    {
-                        gameObject = armatureGo;
-                    }
-                }
-            }
-        }
-
-        private bool TryFindSkeletonFromClip(AnimationClip clip, out GameObject armatureGo)
-        {
-            var path = AssetDatabase.GetAssetPath(ClipTarget.Clip);
-            var owner = AssetDatabase.LoadMainAssetAtPath(path);
-            var importer = AssetImporter.GetAtPath(AssetDatabase.GetAssetPath(owner)) as ModelImporter;
-            if (importer != null && importer.sourceAvatar != null)
-            {
-                var avatarPath = AssetDatabase.GetAssetPath(importer.sourceAvatar);
-                var avatarOwner = AssetDatabase.LoadMainAssetAtPath(avatarPath);
-                if (avatarOwner is GameObject go)
-                {
-                    armatureGo = go;
-                    return IsValidGameObject(go);
-                }
-            }
-
-            armatureGo = null;
-            return false;
+            preview?.Dispose();
         }
 
         public override bool HasPreviewGUI()
@@ -216,27 +85,7 @@ namespace DOTSAnimation.Editor
 
         public override void OnPreviewGUI(Rect r, GUIStyle background)
         {
-            if (skinnedMeshRenderer != null && ClipTarget.Clip != null)
-            {
-                AnimationMode.BeginSampling();
-                AnimationMode.SamplePlayableGraph(playableGraph, 0, sampleNormalizedTime*ClipTarget.Clip.length);
-                AnimationMode.EndSampling();
-                {
-                    skinnedMeshRenderer.BakeMesh(previewMesh);
-                    previewRenderUtility.BeginPreview(r, background);
-
-                    for (var i = 0; i < previewMesh.subMeshCount; i++)
-                    {
-                        previewRenderUtility.DrawMesh(previewMesh, Matrix4x4.identity,
-                            skinnedMeshRenderer.sharedMaterial, i);
-                    }
-
-                    previewRenderUtility.camera.Render();
-
-                    var resultRender = previewRenderUtility.EndPreview();
-                    GUI.DrawTexture(r, resultRender, ScaleMode.StretchToFill, false);
-                }
-            }
+            preview.DrawPreview(r, background);
         }
     }
 }
